+++ src/event_loop.rs
@@ -38,7 +38,7 @@ impl<'a> EventLoop<'a> {
                   render_frame.height = height;
                },
                Event::MouseMoved(x, y) => {
-                  renderer.mouse_moved(x, render_frame.height - y - 1);
+                  renderer.mouse_moved(x, render_frame.height as i32 - y - 1);
                },
                Event::Quit => {
                   quit = true;
